pyproject.toml
[tool.poetry]
name = "littlelemon"
version = "0.1.0"
description = "Website for the Little Lemon distro"
authors = ["Winston Kinoti <prog.kinoti@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.3"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
djangorestframework-xml = "^2.0.0"
drf-spectacular = "^0.26.3"
mysqlclient = "^2.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
ipython = "^8.14.0"
django-browser-reload = "^1.11.0"
django-debug-toolbar = "^4.1.0"
django-fastdev = "^1.7.3"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.7"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
